add validation annotation

+++ umc_study/umc/src/main/java/umc/umc/study/repository/MissionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import umc.umc.study.domain.Member;
 import umc.umc.study.domain.Mission;
 
-public interface MissionRepository extends JpaRepository<Mission, Integer> {
+public interface MissionRepository extends JpaRepository<Mission, Long> {
 }
